# source-mysql-batch: Optimize document serialization