**Description:**

This adds a new helper package `./go/encrow` for serializing tabular data such as we get from SQL queries, and modifies `source-mysql-batch` to use this new package.

The helper package is more efficient than using the standard library `encoding/json` to serialize a `map[string]any` for two reasons:

1. The top level of document serialization is specialized to only work on objects with a specific arity and list of property names. This is represented with a new `encrow.Shape` abstraction.
2. Serializing the individual values of each document is done with the github.com/segmentio/encoding/json package, which is a micro-optimized JSON encoder/decoder package intended as a faster drop-in replacement for the stdlib one.

The specialization alone resulted in about a 3x speedup even while still using the stdlib encoder for each field value. Adding the segmentio JSON encoder provided another 2x speedup on top of that.

Benchmark results for demonstration (measuring single core performance on my laptop):

    $ GOMAXPROCS=1 go test -v ./go/encrow/ -bench=. -benchtime=30s -count=3
    cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
    BenchmarkStandardSerialization
    BenchmarkStandardSerialization      4982           7267985 ns/op
    BenchmarkStandardSerialization      4867           7644533 ns/op
    BenchmarkStandardSerialization      4099           8395409 ns/op
    BenchmarkShapeSerialization
    BenchmarkShapeSerialization        26900           1334689 ns/op
    BenchmarkShapeSerialization        26907           1337353 ns/op
    BenchmarkShapeSerialization        27273           1327145 ns/op

I have written a normal test and a fuzz test which attempt to verify that for every possible object mapping keys to values, the output of the new encoder exactly matches the standard library encoder. This turned up one bug (now fixed) in how an object with zero properties was serialized, and other than that I have seen no failures after a decently large amount of fuzzing.

So as far as I can tell this ought to be a pretty straightforward improvement to throughput. Currently just for `source-mysql-batch` but this same package can be used for all our SQL captures in the future.

**Workflow steps:**

Nothing changes, `source-mysql-batch` captures run up to 6x faster.

**Notes for reviewers:**

I'm not 100% settled on this design, but I was just playing around with approaches to optimizing our JSON serialization of result rows and this seems to work pretty well, as demonstrated in the benchmark results included up above.

It would be even simpler (literally a one-line code change) to just drop in the segmentio JSON encoder instead of the standard library one but skip all the "shape optimization" stuff, however the custom code I wrote for specializing the object shape really does provide a noticeable improvement (about 2x) even on top of that.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/estuary/connectors/1005)
<!-- Reviewable:end -->